pyproject.toml
[project]
name = "agentstr-sdk"
version = "0.1.8"
description = "APython SDK for building decentralized agentic applications on Nostr."
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]
authors = [
    { name = "@HallmarkBitcoin", email = "ehallmark1122@gmail.com" }
]
dependencies = [
    "bolt11>=2.1.1",
    "expiringdict>=1.2.2",
    "mcp[cli]>=1.6.0",
    "pycryptodome==3.10.1",
    "pydantic>=2.11.3",
    "pynostr[websocket-client]>=0.6.2",
    "python-dotenv>=1.1.0",
    "secp256k1>=0.14.0",
    "websockets>=15.0.1",
]
readme = "README.md"
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "furo>=2024.8.6",
    "sphinx>=8.1.3",
    "sphinx-rtd-theme>=3.0.2",
]
examples = [
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.7",
]

[project.urls]
Homepage = "https://github.com/agentstr/agentstr-sdk"
Issues = "https://github.com/agentstr/agentstr-sdk/issues"

[project.optional-dependencies]
rag = [
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.18",
]

